fix load table script

+++ sql/wis/load_table_fileondisk.sh
@@ -1,4 +1,4 @@
-#!/bin/sh
+#!/bin/bash
 
 HOST=localhost
 PORT=2345
@@ -6,8 +6,18 @@ DB=pgraster
 USER=loader
 export PGPASSWORD=ChangeMe
 
-echo 'DROP TABLE IF EXISTS fileondisk;\nCREATE TABLE fileondisk(name text);' > creattable.sql
-find /home/alessard/data/ -type f -iname *.xml -printf "INSERT INTO fileondisk(name) VALUES('%f');\n" >> creattable.sql
+if [[ $# != 2 ]]; then
+    echo "Load in postgresql files located on disk"
+    echo "USAGE: bash ./load_table_fileondisk.sh -[ac](append or create table) [directory]"
+    exit 0
+fi
+
+if [[ "$1" -eq "-c" ]]; then
+    echo 'DROP TABLE IF EXISTS fileondisk;\nCREATE TABLE fileondisk(name text);' > creattable.sql
+    find $2 -type f -iname *.xml -printf "INSERT INTO fileondisk(name) VALUES('%f');\n" >> creattable.sql
+elif if [[ "$1" -eq "-a" ]]; then
+    find $2 -type f -iname *.xml -printf "INSERT INTO fileondisk(name) VALUES('%f');\n" >> creattable.sql
+fi
 
 psql -h $HOST -p $PORT -d $DB -U $USER -f creattable.sql
